\name{dfexplore-package}
\alias{dfexplore-package}
\alias{dfexplore}
\docType{package}
\title{
Data-frame explorer
}
\description{
This package aim to explore quickly a data.frame. For example, find visually NA and classes of data.
}
\details{
\tabular{ll}{
Package: \tab dfexplore\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2013-12-06\cr
License: \tab GPL\cr
}
}
\author{
Joris Muller

Maintainer: Joris Muller <joris.muller@etu.unistra.fr>

Testing: Maël Barthoulot
}

\keyword{ package }
\seealso{

\code{\link[=dfplot]{dfplot}}
}
\examples{
dfplot(example_df)
}
